bond-extractor/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
pytesseract
pypdfium2
python-dateutil
regex
requests
paddlepaddle
paddleocr
transformers
sentence-transformers
torch
sentencepiece
accelerate
protobuf

mistralai
